+++ src/lib/game/mouse.ts
