+++ src/lib/game/piece.ts
