+++ src/lib/game/vec.ts
